feat: set rebase.updateRefs = true
This makes it very quick and easy to rebase multiple branches that were
branched off one another in case one of the branches had changes.

In case of:

main
    \
     A
      \
       B
        \
         C

Now if a change was made to branch A, all that needs doing is to
checkout C and use `git rebase A` for all the intermediate refs to be
updated and keep a linear history.