feat(git): make whitespace visible in diffs
Difftastic, which is used when using `git difft` uses an AST to show
diffs. This means that when a change consists of only whitespace
changes, it will tell the user that there are not syntactic changes.

Just using `git diff` however still works with this new setting and will
happily show whitespace changes.